Messed with build system again to allow us to have different object files per board
This will allow us to do conditional compilation on an object file
depending upon the flags set. For example, if two boards use a button
they can use the common botton interface but in the init function haf an
ifdef to determine which board it is and set the pins appropriately.